Add misc_utils file, moved progressbar function
