+++ recognition/Segmentation-ISICs-ImprovedUnet-s4623205/run_Unet_ISICs.py
@@ -1,26 +1,8 @@
 import matplotlib.pyplot as plt
 import cv2
-import sys
-from math import ceil
 import glob
 import numpy as np
-
-
-def progressbar(current, max_size):
-    """
-    Function for displaying the progress in the console
-
-    Parameters
-    ----------
-    current : integer
-      The current progress count
-    max_size : integer
-      The maximum progress count
-    """
-    sys.stdout.write('\r')
-    progress = ceil((100 / int(max_size)) * current)
-    sys.stdout.write("[%-100s] %d%%" % ('=' * progress, progress))
-    sys.stdout.flush()
+from misc_utils import progressbar
 
 
 def get_min_imageshape(path):
